requirements.txt
apache-beam==2.48.0
apache-flink==1.20.1
apache-flink-libraries==1.20.1
argparse-addons==0.12.0
asttokens==3.0.0
avro-python3==1.10.2
backcall==0.2.0
bitstruct==8.20.0
cantools==39.4.13
certifi==2025.1.31
charset-normalizer==3.4.1
cloudpickle==2.2.1
comm==0.2.2
confluent-kafka==2.8.2
crccheck==1.3.0
crcmod==1.7
debugpy==1.8.12
decorator==5.2.1
dill==0.3.1.1
diskcache==5.6.3
dnspython==2.6.1
docopt==0.6.2
executing==2.2.0
fastavro==1.9.7
fasteners==0.19
find_libpython==0.4.0
grpcio==1.70.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
logger==1.4
matplotlib-inline==0.1.7
msgpack==1.1.0
nest-asyncio==1.6.0
numpy==1.24.4
objsize==0.6.1
orjson==3.10.15
packaging==24.2
pandas==2.0.3
parso==0.8.4
pemja==0.4.1
pexpect==4.9.0
pickleshare==0.7.5
pkg_resources==0.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==4.23.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==11.0.0
pydot==1.4.2
pyflink==1.0
Pygments==2.19.1
pymongo==4.10.1
pyparsing==3.1.4
python-can==4.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
six==1.17.0
stack-data==0.6.3
textparser==0.24.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.20.2
zstandard==0.23.0
